Refactor into model
